Grid
Not finished yet, don't panic! Images need to be somehow resized (thinking about using a Glasspane, I'll tell you why if you're curious) , Saturday I'm going to place the little character tokens on the board, will be really easy since the grid has been mapped properly to a 2D array!! finally going to create a tabbed pane on the bottom panel to add things like the Weapons and Character Pictures instead of being placed on the Output Console (really easy to do and much neater!) Hope this is ok for now.

Also! I personally, cut and created the images for the Rooms, Characters and Weapons! so I will place those in files named accordingly!