Cards now showing for six players
Just added an equal sign to the do...while statement on line 103. Cards were only being allocated to players when the amount of players was less than the max number of players. Changed this to allow the cards to be allocated to players when they were less than or equal to the the max.